Allow setting metadataBase for metadata
