+++ packages/next/src/lib/metadata/generate/meta.tsx
@@ -9,7 +9,7 @@ export function Meta({
   property?: string
   content: string | number | URL | null | undefined
 }): React.ReactElement | null {
-  if (typeof content !== 'undefined' && content !== null) {
+  if (typeof content !== 'undefined' && content != null) {
     return (
       <meta
         {...(name ? { name } : { property })}
