+++ packages/next/src/lib/metadata/resolve-metadata.ts
@@ -3,18 +3,10 @@ import type {
   ResolvedMetadata,
   ResolvingMetadata,
 } from './types/metadata-interface'
-import type { Viewport } from './types/extra-types'
-import type { ResolvedTwitterMetadata } from './types/twitter-types'
-import type {
-  AbsoluteTemplateString,
-  Icon,
-  IconDescriptor,
-  Icons,
-  ResolvedVerification,
-} from './types/metadata-types'
+import type { AbsoluteTemplateString } from './types/metadata-types'
 import { createDefaultMetadata } from './default-metadata'
-import { resolveOpenGraph } from './resolve-opengraph'
-import { mergeTitle } from './resolve-title'
+import { resolveOpenGraph, resolveTwitter } from './resolvers/resolve-opengraph'
+import { mergeTitle } from './resolvers/resolve-title'
 import { resolveAsArrayOrUndefined } from './generate/utils'
 import { isClientReference } from '../../build/is-client-reference'
 import {
@@ -23,214 +15,15 @@ import {
 } from '../../server/lib/app-dir-module'
 import { ComponentsType } from '../../build/webpack/loaders/next-app-loader'
 import { interopDefault } from '../interop-default'
-
-type FieldResolver<Key extends keyof Metadata> = (
-  T: Metadata[Key]
-) => ResolvedMetadata[Key]
-
-const viewPortKeys = {
-  width: 'width',
-  height: 'height',
-  initialScale: 'initial-scale',
-  minimumScale: 'minimum-scale',
-  maximumScale: 'maximum-scale',
-  viewportFit: 'viewport-fit',
-} as const
-
-const resolveViewport: FieldResolver<'viewport'> = (viewport) => {
-  let resolved: ResolvedMetadata['viewport'] = null
-
-  if (typeof viewport === 'string') {
-    resolved = viewport
-  } else if (viewport) {
-    resolved = ''
-    for (const viewportKey_ in viewPortKeys) {
-      const viewportKey = viewportKey_ as keyof Viewport
-      if (viewport[viewportKey]) {
-        if (resolved) resolved += ', '
-        resolved += `${viewPortKeys[viewportKey]}=${viewport[viewportKey]}`
-      }
-    }
-  }
-  return resolved
-}
-
-const VerificationKeys = ['google', 'yahoo', 'yandex', 'me', 'other'] as const
-const resolveVerification: FieldResolver<'verification'> = (verification) => {
-  if (!verification) return null
-  const res = {} as ResolvedVerification
-
-  for (const key of VerificationKeys) {
-    const value = verification[key]
-    if (value) {
-      if (key === 'other') {
-        res.other = {}
-        for (const otherKey in verification.other) {
-          const otherValue = resolveAsArrayOrUndefined(
-            verification.other[otherKey]
-          )
-          if (otherValue) res.other[otherKey] = otherValue
-        }
-      } else res[key] = resolveAsArrayOrUndefined(value) as (string | number)[]
-    }
-  }
-  return res
-}
-
-function isStringOrURL(icon: any): icon is string | URL {
-  return typeof icon === 'string' || icon instanceof URL
-}
-
-function resolveIcon(icon: Icon): IconDescriptor {
-  if (isStringOrURL(icon)) return { url: icon }
-  else if (Array.isArray(icon)) return icon
-  return icon
-}
-
-const IconKeys = ['icon', 'shortcut', 'apple', 'other'] as (keyof Icons)[]
-const TwitterBasicInfoKeys = [
-  'site',
-  'siteId',
-  'creator',
-  'creatorId',
-  'description',
-] as const
-
-const resolveIcons: FieldResolver<'icons'> = (icons) => {
-  if (!icons) {
-    return null
-  }
-
-  const resolved: ResolvedMetadata['icons'] = {}
-  if (Array.isArray(icons)) {
-    resolved.icon = icons.map(resolveIcon).filter(Boolean)
-  } else if (isStringOrURL(icons)) {
-    resolved.icon = [resolveIcon(icons)]
-  } else {
-    for (const key of IconKeys) {
-      const values = resolveAsArrayOrUndefined(icons[key])
-      if (values) resolved[key] = values.map(resolveIcon)
-    }
-  }
-  return resolved
-}
-
-const resolveAppleWebApp: FieldResolver<'appleWebApp'> = (appWebApp) => {
-  if (!appWebApp) return null
-  if (appWebApp === true) {
-    return {
-      capable: true,
-    }
-  }
-
-  const startupImages = resolveAsArrayOrUndefined(appWebApp.startupImage)?.map(
-    (item) => (typeof item === 'string' ? { url: item } : item)
-  )
-
-  return {
-    capable: 'capable' in appWebApp ? !!appWebApp.capable : true,
-    title: appWebApp.title || null,
-    startupImage: startupImages || null,
-    statusBarStyle: appWebApp.statusBarStyle || 'default',
-  }
-}
-
-const resolveTwitter: FieldResolver<'twitter'> = (twitter) => {
-  if (!twitter) return null
-  const resolved = {
-    title: twitter.title,
-  } as ResolvedTwitterMetadata
-  for (const infoKey of TwitterBasicInfoKeys) {
-    resolved[infoKey] = twitter[infoKey] || null
-  }
-  resolved.images = resolveAsArrayOrUndefined(twitter.images)?.map((item) => {
-    if (isStringOrURL(item))
-      return {
-        url: item.toString(),
-      }
-    else {
-      return {
-        url: item.url.toString(),
-        alt: item.alt,
-      }
-    }
-  })
-  if ('card' in twitter) {
-    resolved.card = twitter.card
-    switch (twitter.card) {
-      case 'player': {
-        // @ts-ignore
-        resolved.players = resolveAsArrayOrUndefined(twitter.players) || []
-        break
-      }
-      case 'app': {
-        // @ts-ignore
-        resolved.app = twitter.app || {}
-        break
-      }
-      default:
-        break
-    }
-  } else {
-    resolved.card = 'summary'
-  }
-  return resolved
-}
-
-const resolveAppLinks: FieldResolver<'appLinks'> = (appLinks) => {
-  if (!appLinks) return null
-  for (const key in appLinks) {
-    // @ts-ignore // TODO: type infer
-    appLinks[key] = resolveAsArrayOrUndefined(appLinks[key])
-  }
-  return appLinks as ResolvedMetadata['appLinks']
-}
-
-const robotsKeys = [
-  'noarchive',
-  'nosnippet',
-  'noimageindex',
-  'nocache',
-  'notranslate',
-  'indexifembedded',
-  'nositelinkssearchbox',
-  'unavailable_after',
-  'max-video-preview',
-  'max-image-preview',
-  'max-snippet',
-] as const
-const resolveRobotsValue: (robots: Metadata['robots']) => string | null = (
-  robots
-) => {
-  if (!robots) return null
-  if (typeof robots === 'string') return robots
-
-  const values: string[] = []
-
-  if (robots.index) values.push('index')
-  else if (typeof robots.index === 'boolean') values.push('noindex')
-
-  if (robots.follow) values.push('follow')
-  else if (typeof robots.follow === 'boolean') values.push('nofollow')
-
-  for (const key of robotsKeys) {
-    const value = robots[key]
-    if (typeof value !== 'undefined' && value !== false) {
-      values.push(typeof value === 'boolean' ? key : `${key}:${value}`)
-    }
-  }
-
-  return values.join(', ')
-}
-
-const resolveRobots: FieldResolver<'robots'> = (robots) => {
-  if (!robots) return null
-  return {
-    basic: resolveRobotsValue(robots),
-    googleBot:
-      typeof robots !== 'string' ? resolveRobotsValue(robots.googleBot) : null,
-  }
-}
+import {
+  resolveAlternates,
+  resolveAppleWebApp,
+  resolveAppLinks,
+  resolveRobots,
+  resolveVerification,
+  resolveViewport,
+} from './resolvers/resolve-basics'
+import { resolveIcons } from './resolvers/resolve-icons'
 
 // Merge the source metadata into the resolved target metadata.
 function merge(
@@ -242,6 +35,7 @@ function merge(
     twitter: string | null
   }
 ) {
+  const metadataBase = source.metadataBase || null
   for (const key_ in source) {
     const key = key_ as keyof Metadata
 
@@ -253,19 +47,19 @@ function merge(
         }
         break
       }
+      case 'alternates': {
+        target.alternates = resolveAlternates(source.alternates, metadataBase)
+        break
+      }
       case 'openGraph': {
-        if (typeof source.openGraph !== 'undefined') {
-          target.openGraph = resolveOpenGraph(source.openGraph)
-          if (source.openGraph) {
-            mergeTitle(target.openGraph, templateStrings.openGraph)
-          }
-        } else {
-          target.openGraph = null
+        target.openGraph = resolveOpenGraph(source.openGraph, metadataBase)
+        if (target.openGraph) {
+          mergeTitle(target.openGraph, templateStrings.openGraph)
         }
         break
       }
       case 'twitter': {
-        target.twitter = resolveTwitter(source.twitter)
+        target.twitter = resolveTwitter(source.twitter, metadataBase)
         if (target.twitter) {
           mergeTitle(target.twitter, templateStrings.twitter)
         }
@@ -314,13 +108,15 @@ function merge(
       case 'referrer':
       case 'colorScheme':
       case 'itunes':
-      case 'alternates':
       case 'formatDetection':
         // @ts-ignore TODO: support inferring
         target[key] = source[key] || null
         break
       case 'other':
         target.other = Object.assign({}, target.other, source.other)
+        break
+      case 'metadataBase':
+        target.metadataBase = metadataBase
         break
       default:
         break
