+++ packages/next/src/lib/metadata/resolvers/resolve-url.ts
@@ -1,0 +1,40 @@
+import path from 'path'
+import type { ResolvedMetadata } from '../types/metadata-interface'
+
+export function isStringOrURL(icon: any): icon is string | URL {
+  return typeof icon === 'string' || icon instanceof URL
+}
+
+export function resolveUrlValuesOfObject(
+  obj: Record<string, string | URL | null> | null | undefined,
+  metadataBase: ResolvedMetadata['metadataBase']
+): null | Record<string, URL | null> {
+  if (!obj) return null
+  const result: Record<string, URL | null> = {}
+  for (const [key, value] of Object.entries(obj)) {
+    result[key as keyof typeof obj] = resolveUrl(value, metadataBase)
+  }
+  return result
+}
+
+export function resolveUrl(
+  url: string | URL | null | undefined,
+  metadataBase: URL | null
+): URL | null {
+  if (!url) return null
+  if (url instanceof URL) return url
+
+  try {
+    // If we can construct a URL instance from url, ignore metadataBase
+    const parsedUrl = new URL(url)
+    return parsedUrl
+  } catch (_) {}
+
+  if (!metadataBase) throw new Error('missing metadataBase')
+
+  // Handle relative or absolute paths
+  const basePath = metadataBase.pathname || '/'
+  const joinedPath = path.join(basePath, url)
+
+  return new URL(joinedPath, metadataBase)
+}
