+++ packages/next/src/lib/metadata/types/resolvers.ts
@@ -1,0 +1,9 @@
+import { Metadata, ResolvedMetadata } from './metadata-interface'
+
+export type FieldResolver<Key extends keyof Metadata> = (
+  T: Metadata[Key]
+) => ResolvedMetadata[Key]
+export type FieldResolverWithMetadataBase<Key extends keyof Metadata> = (
+  T: Metadata[Key],
+  metadataBase: ResolvedMetadata['metadataBase']
+) => ResolvedMetadata[Key]
