+++ test/e2e/app-dir/metadata/app/metadata-base/layout.tsx
@@ -1,0 +1,7 @@
+export default function layout({ children }) {
+  return children
+}
+
+export const metadata = {
+  metadataBase: new URL('https://example.com/base'),
+}
