+++ test/e2e/app-dir/metadata/app/metadata-base/url-instance/page.tsx
@@ -1,0 +1,10 @@
+export default function page() {
+  return 'url'
+}
+
+export const metadata = {
+  metadataBase: new URL('https://example.com/foo'),
+  openGraph: {
+    url: new URL('http://https://outerspace.com/huozhi.png'),
+  },
+}
