+++ test/e2e/app-dir/metadata/app/metadata-base/url/page.tsx
@@ -1,0 +1,10 @@
+export default function page() {
+  return 'url'
+}
+
+export const metadata = {
+  metadataBase: new URL('https://bar.example/url'),
+  alternates: {
+    canonical: 'subpath',
+  },
+}
